--- a/src/Compilers/Core/CodeAnalysisTest/StringTableTests.cs
+++ b/src/Compilers/Core/CodeAnalysisTest/StringTableTests.cs
@@ -1,5 +1,6 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+using System;
 using System.Text;
 using Microsoft.CodeAnalysis.Text;
 using Roslyn.Utilities;
@@ -83,7 +84,7 @@ namespace Microsoft.CodeAnalysis.UnitTests
         {
             fixed (byte* ptr = Encoding.ASCII.GetBytes(ascii))
             {
-                var ptrResult = StringTable.TextEqualsASCII(str, ptr, ascii.Length);
+                var ptrResult = StringTable.TextEqualsASCII(str, new ReadOnlySpan<byte>(ptr, ascii.Length));
                 var sbResult = StringTable.TextEquals(str, new StringBuilder(ascii));
                 var substrResult = StringTable.TextEquals(str, "xxx" + ascii + "yyy", 3, ascii.Length);
                 Assert.Equal(substrResult, sbResult);