--- a/src/Compilers/Core/Portable/InternalUtilities/Hash.cs
+++ b/src/Compilers/Core/Portable/InternalUtilities/Hash.cs
@@ -178,16 +178,15 @@ namespace Roslyn.Utilities
         /// See http://en.wikipedia.org/wiki/Fowler%E2%80%93Noll%E2%80%93Vo_hash_function
         /// </summary>
         /// <param name="data">The sequence of bytes that are likely to be ASCII text.</param>
-        /// <param name="length">The length of the sequence.</param>
         /// <param name="isAscii">True if the sequence contains only characters in the ASCII range.</param>
         /// <returns>The FNV-1a hash of <paramref name="data"/></returns>
-        internal static unsafe int GetFNVHashCode(byte* data, int length, out bool isAscii)
+        internal static int GetFNVHashCode(ReadOnlySpan<byte> data, out bool isAscii)
         {
             int hashCode = Hash.FnvOffsetBias;
 
             byte asciiMask = 0;
 
-            for (int i = 0; i < length; i++)
+            for (int i = 0; i < data.Length; i++)
             {
                 byte b = data[i];
                 asciiMask |= b;