--- a/src/Compilers/Core/Portable/InternalUtilities/StringTable.cs
+++ b/src/Compilers/Core/Portable/InternalUtilities/StringTable.cs
@@ -109,6 +109,7 @@ namespace Roslyn.Utilities
 
         internal string Add(char[] chars, int start, int len)
         {
+            var span = chars.AsSpan(start, len);
             var hashCode = Hash.GetFNVHashCode(chars, start, len);
 
             // capture array to avoid extra range checks
@@ -120,7 +121,7 @@ namespace Roslyn.Utilities
             if (text != null && arr[idx].HashCode == hashCode)
             {
                 var result = arr[idx].Text;
-                if (StringTable.TextEquals(result, chars, start, len))
+                if (StringTable.TextEquals(result, span))
                 {
                     return result;
                 }
@@ -294,7 +295,7 @@ namespace Roslyn.Utilities
 
                 if (e != null)
                 {
-                    if (hash == hashCode && TextEquals(e, chars, start, len))
+                    if (hash == hashCode && TextEquals(e, chars.AsSpan(start, len)))
                     {
                         break;
                     }
@@ -349,7 +350,7 @@ namespace Roslyn.Utilities
             return e;
         }
 
-        private static unsafe string FindSharedEntryASCII(int hashCode, byte* asciiChars, int length)
+        private static string FindSharedEntryASCII(int hashCode, ReadOnlySpan<byte> asciiChars)
         {
             var arr = s_sharedTable;
             int idx = SharedIdxFromHash(hashCode);
@@ -364,7 +365,7 @@ namespace Roslyn.Utilities
 
                 if (e != null)
                 {
-                    if (hash == hashCode && TextEqualsASCII(e, asciiChars, length))
+                    if (hash == hashCode && TextEqualsASCII(e, asciiChars))
                     {
                         break;
                     }
@@ -580,29 +581,34 @@ namespace Roslyn.Utilities
             return text;
         }
 
-        internal static unsafe string AddSharedUTF8(byte* bytes, int byteCount)
+        internal static string AddSharedUTF8(ReadOnlySpan<byte> bytes)
         {
             bool isAscii;
-            int hashCode = Hash.GetFNVHashCode(bytes, byteCount, out isAscii);
+            int hashCode = Hash.GetFNVHashCode(bytes, out isAscii);
 
             if (isAscii)
             {
-                string shared = FindSharedEntryASCII(hashCode, bytes, byteCount);
+                string shared = FindSharedEntryASCII(hashCode, bytes);
                 if (shared != null)
                 {
                     return shared;
                 }
             }
 
-            return AddSharedSlow(hashCode, bytes, byteCount, isAscii);
+            return AddSharedSlow(hashCode, bytes, isAscii);
         }
 
-        private static unsafe string AddSharedSlow(int hashCode, byte* utf8Bytes, int byteCount, bool isAscii)
+        private static string AddSharedSlow(int hashCode, ReadOnlySpan<byte> utf8Bytes, bool isAscii)
         {
-            // TODO: This should be Encoding.UTF8.GetString (for better layering) but the unsafe variant isn't portable. 
-            //       The MetadataReader has code to light it up and even fall back to internal String.CreateStringFromEncoding
-            //       on .NET < 4.5.3. Use that instead of copying the light-up code here.
-            string text = System.Reflection.Metadata.MetadataStringDecoder.DefaultUTF8.GetString(utf8Bytes, byteCount);
+            string text;
+
+            unsafe
+            {
+                fixed (byte* bytes = &utf8Bytes.GetPinnableReference())
+                {
+                    text = Encoding.UTF8.GetString(bytes, utf8Bytes.Length);
+                }
+            }
 
             // Don't add non-ascii strings to table. The hashCode we have here is not correct and we won't find them again.
             // Non-ascii in UTF8-encoded parts of metadata (the only use of this at the moment) is assumed to be rare in 
@@ -705,21 +711,21 @@ namespace Roslyn.Utilities
             return true;
         }
 
-        internal static unsafe bool TextEqualsASCII(string text, byte* ascii, int length)
+        internal static bool TextEqualsASCII(string text, ReadOnlySpan<byte> ascii)
         {
 #if DEBUG
-            for (var i = 0; i < length; i++)
+            for (var i = 0; i < ascii.Length; i++)
             {
                 Debug.Assert((ascii[i] & 0x80) == 0, "The byte* input to this method must be valid ASCII.");
             }
 #endif
 
-            if (length != text.Length)
+            if (ascii.Length != text.Length)
             {
                 return false;
             }
 
-            for (var i = 0; i < length; i++)
+            for (var i = 0; i < ascii.Length; i++)
             {
                 if (ascii[i] != text[i])
                 {
@@ -730,25 +736,7 @@ namespace Roslyn.Utilities
             return true;
         }
 
-        internal static bool TextEquals(string array, char[] text, int start, int length)
-        {
-            return array.Length == length && TextEqualsCore(array, text, start);
-        }
-
-        private static bool TextEqualsCore(string array, char[] text, int start)
-        {
-            // use array.Length to eliminate the range check
-            int s = start;
-            for (var i = 0; i < array.Length; i++)
-            {
-                if (array[i] != text[s])
-                {
-                    return false;
-                }
-                s++;
-            }
-
-            return true;
-        }
+        internal static bool TextEquals(string array, ReadOnlySpan<char> text)
+            => text.Equals(array.AsSpan(), StringComparison.Ordinal);
     }
 }