Allows us to replace some unsafe code with Span<T>. I anticipate more complex
uses in the future.

Here are the benchmarks.

**Before**

```
BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17134.345 (1803/April2018Update/Redstone4)
Intel Core i7-7700 CPU 3.60GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.2.200-preview-009431
  [Host] : .NET Core 2.1.6 (CoreCLR 4.6.27019.06, CoreFX 4.6.27019.05), 64bit RyuJIT
  Core   : .NET Core 2.1.6 (CoreCLR 4.6.27019.06, CoreFX 4.6.27019.05), 64bit RyuJIT

Job=Core  Runtime=Core  Server=True

                Method |      Mean |     Error |    StdDev |    Median |     Gen 0 | Allocated |
---------------------- |----------:|----------:|----------:|----------:|----------:|----------:|
               Parsing |  96.86 ms |  2.568 ms |  7.571 ms |  96.89 ms |         - |   4.25 MB |
 CompileMethodsAndEmit | 688.64 ms | 13.659 ms | 38.971 ms | 681.50 ms | 1000.0000 |  43.97 MB |
     SerializeMetadata | 231.56 ms | 10.722 ms | 31.613 ms | 244.91 ms |         - |  18.06 MB |
```

**After**

```
BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17134.345 (1803/April2018Update/Redstone4)
Intel Core i7-7700 CPU 3.60GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.2.200-preview-009431
  [Host] : .NET Core 2.1.6 (CoreCLR 4.6.27019.06, CoreFX 4.6.27019.05), 64bit RyuJIT
  Core   : .NET Core 2.1.6 (CoreCLR 4.6.27019.06, CoreFX 4.6.27019.05), 64bit RyuJIT

Job=Core  Runtime=Core  Server=True

                Method |      Mean |     Error |    StdDev |    Median |     Gen 0 | Allocated |
---------------------- |----------:|----------:|----------:|----------:|----------:|----------:|
               Parsing |  95.71 ms |  2.381 ms |  7.021 ms |  95.44 ms |         - |   3.59 MB |
 CompileMethodsAndEmit | 708.76 ms | 19.100 ms | 55.108 ms | 707.51 ms | 1000.0000 |  44.02 MB |
     SerializeMetadata | 199.04 ms |  8.754 ms | 24.834 ms | 191.37 ms |         - |  18.06 MB |
```

There does not appear to be any statistically significant difference.