@jcouv 2018-11-15
I assume the benchmarks in the description were for a previous iteration with a broader change.